完善管理员对菜单的管理

+++ src/main/java/com/fch/buffetorder/entity/Food.java
@@ -15,6 +15,6 @@ public class Food {
     private String cateName;
     private String foodNote;
     private BigDecimal foodPrice;
-    private BigDecimal foodWeight;
+    private Integer foodWeight;
     private Integer isEnable;
 }
